Moved class-terminus.php to Terminus.php. Moved old Terminus.php to boot-cl.php
