+++ php/boot-phar.php
@@ -1,4 +1,4 @@
 <?php
 
 define('TERMINUS_ROOT', 'phar://terminus.phar');
-include TERMINUS_ROOT . '/php/terminus.php';
+include TERMINUS_ROOT . '/php/boot-cl.php';
